Trying to fix problems with symlinks on Linux, but not quite there yet. Some uses of QDir::entryList removed (#265)

+++ file-commander-core/src/fasthash.c
@@ -33,7 +33,7 @@
 			(h) *= 0x2127599bf4325c37ULL;	\
 			(h) ^= (h) >> 47; }
 
-uint64_t fasthash64(const void *buf, size_t len, uint64_t seed)
+uint64_t fasthash64(const void *buf, uint64_t len, uint64_t seed)
 {
 	const uint64_t    m = 0x880355f21e6d1965ULL;
 	const uint64_t *pos = (const uint64_t *)buf;
