+++ file-commander-core/src/fasthash.h
@@ -47,7 +47,7 @@ extern "C" {
  * @len:  data size
  * @seed: the seed
  */
-	uint64_t fasthash64(const void *buf, size_t len, uint64_t seed);
+	uint64_t fasthash64(const void *buf, uint64_t len, uint64_t seed);
 
 #ifdef __cplusplus
 }
